WIP break/continue
Trying to rewrite the loop body in a style that clojure likes (where
any logic branch that has a break/continue, we can combine the other
branch with whatever follows).

Seems to be mostly on the right track. Currently the issue is that if
an inner branch inserts a continue, the outer branch can't because
then the inner isn't in tail position. I think maybe when we have to
recursively call the shift function, we shouldn't insert any continues.